requirements.txt
redis
kombu
msal
toml
gevent
celery
gunicorn
sqlparse
requests
setuptools
django-redis
django-filter
python-dateutil
psycopg[binary]
django-rest-knox
django-timescaledb
django-celery-beat
django-cors-headers
django-celery-results
langchain_community
langchain
langchain-openai
langchain-aws
langfuse
langgraph